requirements.txt
python>=3.7.5
pytorch>=1.3.1
transformers>=2.6.0
numpy>=1.17.4
scikit-learn>=0.22.1
matplotlib>=3.1.3
ray>=0.8.2
